+++ ChestsAnywhere/Properties/AssemblyInfo.cs
@@ -1,7 +1,5 @@
 using System.Reflection;
-using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.ChestsAnywhere")]
-[assembly: Guid("5b924f11-db83-424f-894f-6783423559e1")]
-[assembly: AssemblyVersion("1.12.3")]
-[assembly: AssemblyFileVersion("1.12.3")]
+[assembly: AssemblyVersion("1.12.4")]
+[assembly: AssemblyFileVersion("1.12.4")]