+++ DataMaps/Properties/AssemblyInfo.cs
@@ -1,7 +1,5 @@
 using System.Reflection;
-using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.DataMaps")]
-[assembly: Guid("cdeef8af-15cb-4955-8163-e446f212c2f4")]
-[assembly: AssemblyVersion("1.2.0")]
-[assembly: AssemblyFileVersion("1.2.0")]
+[assembly: AssemblyVersion("1.3.0")]
+[assembly: AssemblyFileVersion("1.3.0")]