+++ LookupAnything/Properties/AssemblyInfo.cs
@@ -1,7 +1,5 @@
 using System.Reflection;
-using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.LookupAnything")]
-[assembly: Guid("2a9c0811-b2d2-474a-a84c-ddeb16206a62")]
-[assembly: AssemblyVersion("1.17.0.0")]
-[assembly: AssemblyFileVersion("1.17.0.0")]
+[assembly: AssemblyVersion("1.18.0")]
+[assembly: AssemblyFileVersion("1.18.0")]