+++ SkipIntro/Properties/AssemblyInfo.cs
@@ -1,7 +1,5 @@
 using System.Reflection;
-using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.SkipIntro")]
-[assembly: Guid("56fb522e-923d-4582-ad3d-8ec80f09237a")]
-[assembly: AssemblyVersion("1.7.1.0")]
-[assembly: AssemblyFileVersion("1.7.1.0")]
+[assembly: AssemblyVersion("1.7.2")]
+[assembly: AssemblyFileVersion("1.7.2")]