+++ LookupAnything/release-notes.md
@@ -4,6 +4,7 @@
 ## 1.17.1
 * Updated to SMAPI 2.4.
 * Added support for custom machines and objects from Custom Farming Redux.
+* Fixed debug key working when a menu is open.
 * Fixed typo in debug interface.
 * Improved German translations. (Thanks to Husky110!)