+++ LookupAnything/release-notes.md
@@ -3,7 +3,7 @@
 # Release notes
 ## 1.17.1
 * Updated to SMAPI 2.4.
-* Added support for custom machines and objects from Custom Farming Redux.
+* Added support for custom machines and objects from Custom Farming Redux 2.3.6+.
 * Fixed debug key working when a menu is open.
 * Fixed typo in debug interface.
 * Improved German translations. (Thanks to Husky110!)